Add controls

+++ app/android/app/src/main/java/com/example/arken/model/Alert.kt
@@ -5,6 +5,6 @@ import java.io.Serializable
 class Alert (
     val _id: String? = null,
     val currency: String? = null,
-    val rate: Int? = null,
+    val rate: Double? = null,
     val compare: String? = null
 ) : Serializable
